pyproject.toml
[tool.poetry]
name = "eufm-assistant"
version = "0.1.0"
description = "EUFM Assistant for Horizon Europe proposals"
authors = ["AI Task Force"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
flask = "*"
openai = "*"
google-generativeai = "*"
pyyaml = "*"
python-dotenv = "*"
markdown = "*"
python-docx = "*"
pydantic = "*"
pydantic-settings = "*"
sqlalchemy = "*"
alembic = "*"
requests = "*"
aiohttp = "*"
python-json-logger = "*"
celery = "*"
redis = "*"
psycopg2-binary = "*"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-flask = "*"
pytest-asyncio = "*"
pytest-cov = "*"
black = "*"
isort = "*"
flake8 = "*"
mypy = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"